refactor(stdlib): split web streams module
