+++ crates/perry-stdlib/src/streams.rs
@@ -22,10 +22,6 @@
 //! Stubs: BYOB readers and full custom `QueuingStrategy` size accounting —
 //! see the inline comment on each site.
 
-use flate2::read::{
-    DeflateDecoder, DeflateEncoder, GzDecoder, GzEncoder, ZlibDecoder, ZlibEncoder,
-};
-use flate2::Compression;
 use perry_runtime::{
     js_array_alloc, js_array_push, js_closure_call0, js_closure_call1, js_closure_call2,
     js_nanbox_get_pointer, js_object_alloc, js_object_get_field_by_name, js_object_set_field,
@@ -33,12 +29,41 @@ use perry_runtime::{
     js_promise_resolve, js_string_from_bytes, ClosureHeader, JSValue, ObjectHeader, Promise,
 };
 use std::collections::{HashMap, VecDeque};
-use std::io::Read;
 use std::os::raw::c_int;
 use std::sync::Mutex;
 
 mod pipe;
+mod subclass;
+#[cfg(test)]
+mod tests;
+mod transform;
+mod writable;
+
 use self::pipe::js_readable_stream_pipe_to;
+use self::subclass::{box_promise, js_stream_unwrap_handle};
+pub(crate) use self::subclass::{dispatch_stream_method, dispatch_stream_property};
+pub use self::subclass::{
+    drain_readable_into_bytes, js_readable_stream_subclass_init, js_stream_handle_is_registered,
+    js_stream_handle_kind, js_transform_stream_subclass_init, js_writable_stream_subclass_init,
+};
+pub use self::transform::{
+    js_stream_web_compression_stream_new, js_stream_web_decompression_stream_new,
+    js_stream_web_text_decoder_stream_new, js_stream_web_text_encoder_stream_new,
+    js_streams_throw_byob_not_implemented, js_streams_throw_byte_length_not_implemented,
+    js_transform_stream_new, js_transform_stream_new_from_transformer_object,
+    js_transform_stream_readable, js_transform_stream_writable,
+};
+use self::transform::{
+    transform_close, transform_writable_for_readable, transform_write, TRANSFORM_PAIRS,
+};
+pub use self::writable::{
+    js_writable_stream_abort, js_writable_stream_close, js_writable_stream_get_writer,
+    js_writable_stream_locked, js_writable_stream_new, js_writable_stream_new_from_sink_object,
+    js_writable_stream_new_with_sink_type, js_writable_stream_throw_invalid_sink, js_writer_abort,
+    js_writer_close, js_writer_closed, js_writer_desired_size, js_writer_ready,
+    js_writer_release_lock, js_writer_write,
+};
+use self::writable::{js_writable_stream_abort_inner, writable_stream_write};
 
 const TAG_UNDEFINED: u64 = 0x7FFC_0000_0000_0001;
 const TAG_NULL: u64 = 0x7FFC_0000_0000_0002;
@@ -1856,1822 +1881,4 @@ pub unsafe extern "C" fn js_readable_stream_pipe_through(
         f64::from_bits(TAG_UNDEFINED),
     );
     transform_readable_handle
-}
-
-// ─────────────────────────────────────────────────────────────────────
-// WritableStream FFI
-// ─────────────────────────────────────────────────────────────────────
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_new(
-    start_bits: f64,
-    write_bits: f64,
-    close_bits: f64,
-    abort_bits: f64,
-    hwm: f64,
-) -> f64 {
-    js_writable_stream_new_with_sink_type(
-        start_bits,
-        write_bits,
-        close_bits,
-        abort_bits,
-        hwm,
-        f64::from_bits(TAG_UNDEFINED),
-    )
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_new_with_sink_type(
-    start_bits: f64,
-    write_bits: f64,
-    close_bits: f64,
-    abort_bits: f64,
-    hwm: f64,
-    sink_type: f64,
-) -> f64 {
-    if sink_type.to_bits() != TAG_UNDEFINED {
-        throw_range_error_with_code(
-            "The argument 'type' is invalid. Received a non-undefined value",
-            "ERR_INVALID_ARG_VALUE",
-        );
-    }
-
-    ensure_gc_registered();
-    let id = alloc_writable(
-        closure_from_bits(write_bits.to_bits()),
-        closure_from_bits(close_bits.to_bits()),
-        closure_from_bits(abort_bits.to_bits()),
-        hwm,
-    );
-    // #1545: WritableStream `start(controller)` fires synchronously at
-    // construction (before any write), matching the WHATWG order
-    // start → write → close. The controller arg is the stream handle.
-    let start_cb = closure_from_bits(start_bits.to_bits());
-    if start_cb != 0 {
-        js_closure_call1(start_cb as *const ClosureHeader, id as f64);
-    }
-    id as f64
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_new_from_sink_object(sink: f64, hwm: f64) -> f64 {
-    ensure_gc_registered();
-    let sink_type = stream_object_field(sink, b"type");
-    if sink_type.to_bits() != TAG_UNDEFINED {
-        throw_range_error_with_code(
-            "The argument 'type' is invalid. Received a non-undefined value",
-            "ERR_INVALID_ARG_VALUE",
-        );
-    }
-
-    let id = alloc_writable(
-        stream_object_closure(sink, b"write"),
-        stream_object_closure(sink, b"close"),
-        stream_object_closure(sink, b"abort"),
-        hwm,
-    );
-    let start_cb = stream_object_closure(sink, b"start");
-    if start_cb != 0 {
-        js_closure_call1(start_cb as *const ClosureHeader, id as f64);
-    }
-    id as f64
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_throw_invalid_sink() -> f64 {
-    throw_invalid_arg_type("The \"sink\" argument must be of type object")
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_get_writer(stream_handle: f64) -> f64 {
-    ensure_gc_registered();
-    let id = stream_handle as usize;
-    let mut g = WRITABLE_STREAMS.lock().unwrap();
-    let s = match g.get_mut(&id) {
-        Some(s) => s,
-        None => return f64::from_bits(TAG_UNDEFINED),
-    };
-    if s.writer_handle.is_some() {
-        drop(g);
-        throw_type_error("WritableStream is locked");
-    }
-    let writer_id = next_id(&NEXT_STREAM_ID);
-    s.writer_handle = Some(writer_id);
-    let closed_p = s.closed_promise;
-    let ready_p = s.ready_promise;
-    drop(g);
-    WRITERS.lock().unwrap().insert(
-        writer_id,
-        WriterData {
-            stream_handle: id,
-            locked: true,
-            closed_promise: closed_p,
-            ready_promise: ready_p,
-        },
-    );
-    writer_id as f64
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_locked(stream_handle: f64) -> f64 {
-    let id = stream_handle as usize;
-    let g = WRITABLE_STREAMS.lock().unwrap();
-    let locked = g
-        .get(&id)
-        .map(|s| s.writer_handle.is_some())
-        .unwrap_or(false);
-    f64::from_bits(if locked { TAG_TRUE } else { TAG_FALSE })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_close(stream_handle: f64) -> *mut Promise {
-    let promise = js_promise_new();
-    let id = stream_handle as usize;
-    let start_close = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&id) {
-            Some(s) => match s.state {
-                WritableState::Writable => {
-                    s.state = WritableState::Closing;
-                    s.close_request_promise = promise;
-                    !s.in_flight && s.write_queue.is_empty()
-                }
-                WritableState::Closing => {
-                    if !s.close_request_promise.is_null() {
-                        return s.close_request_promise;
-                    }
-                    s.close_request_promise = promise;
-                    !s.in_flight && s.write_queue.is_empty()
-                }
-                WritableState::Closed => {
-                    js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-                    return promise;
-                }
-                WritableState::Errored => {
-                    js_promise_reject(promise, f64::from_bits(s.error_value));
-                    return promise;
-                }
-            },
-            None => {
-                js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-                return promise;
-            }
-        }
-    };
-    if start_close {
-        finish_writable_close(id);
-    }
-    promise
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_abort(stream_handle: f64, reason: f64) -> *mut Promise {
-    js_writable_stream_abort_inner(stream_handle, reason, false)
-}
-
-unsafe fn js_writable_stream_abort_inner(
-    stream_handle: f64,
-    reason: f64,
-    allow_locked: bool,
-) -> *mut Promise {
-    let promise = js_promise_new();
-    let id = stream_handle as usize;
-    let reason_bits = reason.to_bits();
-    let mut locked_reject = false;
-    let (cb, closed_p, close_request) = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&id) {
-            Some(s) => {
-                if !allow_locked && s.writer_handle.is_some() {
-                    locked_reject = true;
-                    (0, std::ptr::null_mut(), std::ptr::null_mut())
-                } else {
-                    s.state = WritableState::Errored;
-                    s.error_value = reason_bits;
-                    let close_request = s.close_request_promise;
-                    s.close_request_promise = std::ptr::null_mut();
-                    s.close_started = false;
-                    (s.abort_cb, s.closed_promise, close_request)
-                }
-            }
-            None => (0, std::ptr::null_mut(), std::ptr::null_mut()),
-        }
-    };
-    if locked_reject {
-        reject_type_error(promise, "Invalid state: WritableStream is locked");
-        return promise;
-    }
-    if cb != 0 {
-        js_closure_call1(cb as *const ClosureHeader, reason);
-    }
-    if !closed_p.is_null() {
-        js_promise_reject(closed_p, reason);
-    }
-    if !close_request.is_null() {
-        js_promise_reject(close_request, reason);
-    }
-    js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-    promise
-}
-
-// ─────────────────────────────────────────────────────────────────────
-// WritableStreamDefaultWriter FFI
-// ─────────────────────────────────────────────────────────────────────
-
-fn writable_desired_size(s: &WritableStreamData) -> f64 {
-    s.high_water_mark - if s.in_flight { 1.0 } else { 0.0 } - s.write_queue.len() as f64
-}
-
-fn sync_writer_ready_promise(stream_id: usize, writer_id: usize, ready: *mut Promise) {
-    if let Some(w) = WRITERS.lock().unwrap().get_mut(&writer_id) {
-        if w.stream_handle == stream_id {
-            w.ready_promise = ready;
-        }
-    }
-}
-
-pub(super) unsafe fn install_writable_backpressure_ready(stream_id: usize, writer_id: usize) {
-    let ready = js_promise_new();
-    if let Some(s) = WRITABLE_STREAMS.lock().unwrap().get_mut(&stream_id) {
-        s.ready_promise = ready;
-    }
-    sync_writer_ready_promise(stream_id, writer_id, ready);
-}
-
-fn writable_capture_usize(closure: *const ClosureHeader, idx: u32) -> usize {
-    let bits = perry_runtime::closure::js_closure_get_capture_ptr(closure, idx) as u64;
-    f64::from_bits(bits) as usize
-}
-
-fn writable_capture_promise(closure: *const ClosureHeader, idx: u32) -> *mut Promise {
-    perry_runtime::closure::js_closure_get_capture_ptr(closure, idx) as *mut Promise
-}
-
-extern "C" fn writable_write_start_microtask(closure: *const ClosureHeader) -> f64 {
-    unsafe {
-        let stream_id = writable_capture_usize(closure, 0);
-        let writer_id = writable_capture_usize(closure, 1);
-        let cb = perry_runtime::closure::js_closure_get_capture_ptr(closure, 2);
-        let chunk_bits = perry_runtime::closure::js_closure_get_capture_ptr(closure, 3) as u64;
-        let write_promise = writable_capture_promise(closure, 4);
-        run_writable_write(
-            stream_id,
-            writer_id,
-            cb,
-            f64::from_bits(chunk_bits),
-            write_promise,
-        );
-    }
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-extern "C" fn writable_write_fulfilled(closure: *const ClosureHeader, _value: f64) -> f64 {
-    unsafe {
-        let stream_id = writable_capture_usize(closure, 0);
-        let writer_id = writable_capture_usize(closure, 1);
-        let write_promise = writable_capture_promise(closure, 2);
-        finish_writable_write_success(stream_id, writer_id, write_promise);
-    }
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-extern "C" fn writable_write_rejected(closure: *const ClosureHeader, reason: f64) -> f64 {
-    unsafe {
-        let stream_id = writable_capture_usize(closure, 0);
-        let write_promise = writable_capture_promise(closure, 2);
-        finish_writable_write_error(stream_id, write_promise, reason);
-    }
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-unsafe fn attach_writable_write_handlers(
-    stream_id: usize,
-    writer_id: usize,
-    write_promise: *mut Promise,
-    sink_promise: *mut Promise,
-) {
-    let fulfilled_fn = writable_write_fulfilled as *const u8;
-    let rejected_fn = writable_write_rejected as *const u8;
-    perry_runtime::closure::js_register_closure_arity(fulfilled_fn, 1);
-    perry_runtime::closure::js_register_closure_arity(rejected_fn, 1);
-
-    let on_fulfilled = perry_runtime::closure::js_closure_alloc(fulfilled_fn, 3);
-    perry_runtime::closure::js_closure_set_capture_ptr(
-        on_fulfilled,
-        0,
-        (stream_id as f64).to_bits() as i64,
-    );
-    perry_runtime::closure::js_closure_set_capture_ptr(
-        on_fulfilled,
-        1,
-        (writer_id as f64).to_bits() as i64,
-    );
-    perry_runtime::closure::js_closure_set_capture_ptr(on_fulfilled, 2, write_promise as i64);
-
-    let on_rejected = perry_runtime::closure::js_closure_alloc(rejected_fn, 3);
-    perry_runtime::closure::js_closure_set_capture_ptr(
-        on_rejected,
-        0,
-        (stream_id as f64).to_bits() as i64,
-    );
-    perry_runtime::closure::js_closure_set_capture_ptr(
-        on_rejected,
-        1,
-        (writer_id as f64).to_bits() as i64,
-    );
-    perry_runtime::closure::js_closure_set_capture_ptr(on_rejected, 2, write_promise as i64);
-
-    let _ = perry_runtime::promise::js_promise_then(sink_promise, on_fulfilled, on_rejected);
-}
-
-unsafe fn try_call_writable_write(cb: i64, chunk: f64) -> Result<f64, u64> {
-    let trap_buf = perry_runtime::exception::js_try_push();
-    let jumped = perry_runtime::ffi::setjmp::setjmp(trap_buf as *mut c_int);
-    if jumped == 0 {
-        let result = js_closure_call1(cb as *const ClosureHeader, chunk);
-        perry_runtime::exception::js_try_end();
-        Ok(result)
-    } else {
-        let err = perry_runtime::exception::js_get_exception();
-        perry_runtime::exception::js_clear_exception();
-        perry_runtime::exception::js_try_end();
-        Err(err.to_bits())
-    }
-}
-
-unsafe fn try_call_writable_close(cb: i64) -> Result<f64, u64> {
-    let trap_buf = perry_runtime::exception::js_try_push();
-    let jumped = perry_runtime::ffi::setjmp::setjmp(trap_buf as *mut c_int);
-    if jumped == 0 {
-        let result = js_closure_call0(cb as *const ClosureHeader);
-        perry_runtime::exception::js_try_end();
-        Ok(result)
-    } else {
-        let err = perry_runtime::exception::js_get_exception();
-        perry_runtime::exception::js_clear_exception();
-        perry_runtime::exception::js_try_end();
-        Err(err.to_bits())
-    }
-}
-
-extern "C" fn writable_close_fulfilled(closure: *const ClosureHeader, _value: f64) -> f64 {
-    unsafe {
-        let stream_id = writable_capture_usize(closure, 0);
-        let close_promise = writable_capture_promise(closure, 1);
-        finish_writable_close_success(stream_id, close_promise);
-    }
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-extern "C" fn writable_close_rejected(closure: *const ClosureHeader, reason: f64) -> f64 {
-    unsafe {
-        let stream_id = writable_capture_usize(closure, 0);
-        let close_promise = writable_capture_promise(closure, 1);
-        finish_writable_close_error(stream_id, close_promise, reason);
-    }
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-unsafe fn attach_writable_close_handlers(
-    stream_id: usize,
-    close_promise: *mut Promise,
-    sink_promise: *mut Promise,
-) {
-    let fulfilled_fn = writable_close_fulfilled as *const u8;
-    let rejected_fn = writable_close_rejected as *const u8;
-    perry_runtime::closure::js_register_closure_arity(fulfilled_fn, 1);
-    perry_runtime::closure::js_register_closure_arity(rejected_fn, 1);
-
-    let on_fulfilled = perry_runtime::closure::js_closure_alloc(fulfilled_fn, 2);
-    perry_runtime::closure::js_closure_set_capture_ptr(
-        on_fulfilled,
-        0,
-        (stream_id as f64).to_bits() as i64,
-    );
-    perry_runtime::closure::js_closure_set_capture_ptr(on_fulfilled, 1, close_promise as i64);
-
-    let on_rejected = perry_runtime::closure::js_closure_alloc(rejected_fn, 2);
-    perry_runtime::closure::js_closure_set_capture_ptr(
-        on_rejected,
-        0,
-        (stream_id as f64).to_bits() as i64,
-    );
-    perry_runtime::closure::js_closure_set_capture_ptr(on_rejected, 1, close_promise as i64);
-
-    let _ = perry_runtime::promise::js_promise_then(sink_promise, on_fulfilled, on_rejected);
-}
-
-unsafe fn finish_writable_close(stream_id: usize) {
-    let (cb, close_promise) = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&stream_id) {
-            Some(s) if s.state == WritableState::Closing => {
-                if s.in_flight || !s.write_queue.is_empty() || s.close_started {
-                    return;
-                }
-                s.close_started = true;
-                (s.close_cb, s.close_request_promise)
-            }
-            _ => return,
-        }
-    };
-
-    if cb == 0 {
-        finish_writable_close_success(stream_id, close_promise);
-        return;
-    }
-    let result = match try_call_writable_close(cb) {
-        Ok(result) => result,
-        Err(reason) => {
-            finish_writable_close_error(stream_id, close_promise, f64::from_bits(reason));
-            return;
-        }
-    };
-    let sink_promise = perry_runtime::promise::js_promise_resolved(result);
-    if !sink_promise.is_null() {
-        attach_writable_close_handlers(stream_id, close_promise, sink_promise);
-        return;
-    }
-    finish_writable_close_success(stream_id, close_promise);
-}
-
-unsafe fn finish_writable_close_success(stream_id: usize, promise: *mut Promise) {
-    let (ready, closed) = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&stream_id) {
-            Some(s) => {
-                s.state = WritableState::Closed;
-                s.close_request_promise = std::ptr::null_mut();
-                s.close_started = false;
-                (s.ready_promise, s.closed_promise)
-            }
-            None => (std::ptr::null_mut(), std::ptr::null_mut()),
-        }
-    };
-    if !promise.is_null() {
-        js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-    }
-    if !ready.is_null() {
-        js_promise_resolve(ready, f64::from_bits(TAG_UNDEFINED));
-    }
-    if !closed.is_null() {
-        js_promise_resolve(closed, f64::from_bits(TAG_UNDEFINED));
-    }
-}
-
-unsafe fn finish_writable_close_error(stream_id: usize, promise: *mut Promise, reason: f64) {
-    let (ready, closed) = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&stream_id) {
-            Some(s) => {
-                s.state = WritableState::Errored;
-                s.error_value = reason.to_bits();
-                s.close_request_promise = std::ptr::null_mut();
-                s.close_started = false;
-                (s.ready_promise, s.closed_promise)
-            }
-            None => (std::ptr::null_mut(), std::ptr::null_mut()),
-        }
-    };
-    if !promise.is_null() {
-        js_promise_reject(promise, reason);
-    }
-    if !ready.is_null() {
-        js_promise_reject(ready, reason);
-    }
-    if !closed.is_null() {
-        js_promise_reject(closed, reason);
-    }
-}
-
-pub(super) unsafe fn run_writable_write(
-    stream_id: usize,
-    writer_id: usize,
-    cb: i64,
-    chunk: f64,
-    promise: *mut Promise,
-) {
-    if cb == 0 {
-        finish_writable_write_success(stream_id, writer_id, promise);
-        return;
-    }
-    let result = match try_call_writable_write(cb, chunk) {
-        Ok(result) => result,
-        Err(reason) => {
-            finish_writable_write_error(stream_id, promise, f64::from_bits(reason));
-            return;
-        }
-    };
-    let sink_promise = perry_runtime::promise::js_promise_resolved(result);
-    if !sink_promise.is_null() {
-        attach_writable_write_handlers(stream_id, writer_id, promise, sink_promise);
-        return;
-    }
-    finish_writable_write_success(stream_id, writer_id, promise);
-}
-
-pub(super) unsafe fn writable_stream_write(
-    stream_id: usize,
-    writer_id: usize,
-    chunk: f64,
-) -> *mut Promise {
-    if TRANSFORM_PAIRS.lock().unwrap().contains_key(&stream_id) {
-        return transform_write(stream_id, chunk);
-    }
-    let promise = js_promise_new();
-    let mut start_write = None;
-    let needs_pending_ready;
-    {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        let s = match g.get_mut(&stream_id) {
-            Some(s) if s.state == WritableState::Writable => s,
-            Some(s) if s.state == WritableState::Errored => {
-                let e = s.error_value;
-                js_promise_reject(promise, f64::from_bits(e));
-                return promise;
-            }
-            _ => {
-                let err = make_error_with_message("Stream is closed or closing");
-                js_promise_reject(promise, f64::from_bits(err));
-                return promise;
-            }
-        };
-        let before = writable_desired_size(s);
-        if s.in_flight {
-            s.write_queue.push_back((chunk.to_bits(), promise));
-        } else {
-            s.in_flight = true;
-            start_write = Some((s.write_cb, chunk, promise));
-        }
-        let after = writable_desired_size(s);
-        needs_pending_ready = before > 0.0 && after <= 0.0;
-    }
-    if needs_pending_ready {
-        install_writable_backpressure_ready(stream_id, writer_id);
-    }
-    if let Some((cb, chunk, write_promise)) = start_write {
-        let start_fn = writable_write_start_microtask as *const u8;
-        perry_runtime::closure::js_register_closure_arity(start_fn, 0);
-        let start = perry_runtime::closure::js_closure_alloc(start_fn, 5);
-        perry_runtime::closure::js_closure_set_capture_ptr(
-            start,
-            0,
-            (stream_id as f64).to_bits() as i64,
-        );
-        perry_runtime::closure::js_closure_set_capture_ptr(
-            start,
-            1,
-            (writer_id as f64).to_bits() as i64,
-        );
-        perry_runtime::closure::js_closure_set_capture_ptr(start, 2, cb);
-        perry_runtime::closure::js_closure_set_capture_ptr(start, 3, chunk.to_bits() as i64);
-        perry_runtime::closure::js_closure_set_capture_ptr(start, 4, write_promise as i64);
-        perry_runtime::builtins::js_queue_microtask(start as i64);
-    }
-    promise
-}
-
-unsafe fn finish_writable_write_success(stream_id: usize, writer_id: usize, promise: *mut Promise) {
-    if !promise.is_null() {
-        js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-    }
-
-    let (next, ready, close_now) = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&stream_id) {
-            Some(s) => {
-                s.in_flight = false;
-                let next =
-                    if s.state == WritableState::Writable || s.state == WritableState::Closing {
-                        s.write_queue.pop_front().map(|(chunk, p)| {
-                            s.in_flight = true;
-                            (s.write_cb, f64::from_bits(chunk), p)
-                        })
-                    } else {
-                        None
-                    };
-                let ready = if s.state == WritableState::Writable && writable_desired_size(s) > 0.0
-                {
-                    s.ready_promise
-                } else {
-                    std::ptr::null_mut()
-                };
-                let close_now = s.state == WritableState::Closing
-                    && next.is_none()
-                    && !s.in_flight
-                    && s.write_queue.is_empty();
-                (next, ready, close_now)
-            }
-            None => (None, std::ptr::null_mut(), false),
-        }
-    };
-
-    if !ready.is_null() {
-        js_promise_resolve(ready, f64::from_bits(TAG_UNDEFINED));
-    }
-    if let Some((cb, chunk, queued_promise)) = next {
-        run_writable_write(stream_id, writer_id, cb, chunk, queued_promise);
-    } else if close_now {
-        finish_writable_close(stream_id);
-    }
-}
-
-unsafe fn finish_writable_write_error(stream_id: usize, promise: *mut Promise, reason: f64) {
-    let (ready, closed, close_request, queued) = {
-        let mut g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&stream_id) {
-            Some(s) => {
-                s.in_flight = false;
-                s.state = WritableState::Errored;
-                s.error_value = reason.to_bits();
-                let close_request = s.close_request_promise;
-                s.close_request_promise = std::ptr::null_mut();
-                s.close_started = false;
-                let queued: Vec<*mut Promise> = s.write_queue.drain(..).map(|(_, p)| p).collect();
-                (s.ready_promise, s.closed_promise, close_request, queued)
-            }
-            None => (
-                std::ptr::null_mut(),
-                std::ptr::null_mut(),
-                std::ptr::null_mut(),
-                Vec::new(),
-            ),
-        }
-    };
-
-    if !promise.is_null() {
-        js_promise_reject(promise, reason);
-    }
-    for queued_promise in queued {
-        if !queued_promise.is_null() {
-            js_promise_reject(queued_promise, reason);
-        }
-    }
-    if !ready.is_null() {
-        js_promise_reject(ready, reason);
-    }
-    if !close_request.is_null() {
-        js_promise_reject(close_request, reason);
-    }
-    if !closed.is_null() {
-        js_promise_reject(closed, reason);
-    }
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_write(writer_handle: f64, chunk: f64) -> *mut Promise {
-    let writer_id = writer_handle as usize;
-    let stream_id = match WRITERS.lock().unwrap().get(&writer_id) {
-        Some(w) if w.locked => w.stream_handle,
-        _ => {
-            let promise = js_promise_new();
-            let err = make_error_with_message("Writer is no longer locked to a stream");
-            js_promise_reject(promise, f64::from_bits(err));
-            return promise;
-        }
-    };
-    writable_stream_write(stream_id, writer_id, chunk)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_desired_size(writer_handle: f64) -> f64 {
-    let writer_id = writer_handle as usize;
-    let stream_id = match WRITERS.lock().unwrap().get(&writer_id) {
-        Some(w) => w.stream_handle,
-        None => return 0.0,
-    };
-    let g = WRITABLE_STREAMS.lock().unwrap();
-    match g.get(&stream_id) {
-        Some(s) if s.state == WritableState::Writable => writable_desired_size(s),
-        Some(s) if s.state == WritableState::Errored => f64::NAN,
-        _ => 0.0,
-    }
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_close(writer_handle: f64) -> *mut Promise {
-    let writer_id = writer_handle as usize;
-    let stream_id = match WRITERS.lock().unwrap().get(&writer_id) {
-        Some(w) => w.stream_handle,
-        None => {
-            let p = js_promise_new();
-            js_promise_resolve(p, f64::from_bits(TAG_UNDEFINED));
-            return p;
-        }
-    };
-    if TRANSFORM_PAIRS.lock().unwrap().contains_key(&stream_id) {
-        return transform_close(stream_id);
-    }
-    js_writable_stream_close(stream_id as f64)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_abort(writer_handle: f64, reason: f64) -> *mut Promise {
-    let writer_id = writer_handle as usize;
-    let stream_id = match WRITERS.lock().unwrap().get(&writer_id) {
-        Some(w) => w.stream_handle,
-        None => {
-            let p = js_promise_new();
-            js_promise_resolve(p, f64::from_bits(TAG_UNDEFINED));
-            return p;
-        }
-    };
-    js_writable_stream_abort_inner(stream_id as f64, reason, true)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_release_lock(writer_handle: f64) -> f64 {
-    let writer_id = writer_handle as usize;
-    let stream_id = {
-        let mut g = WRITERS.lock().unwrap();
-        match g.get_mut(&writer_id) {
-            Some(w) => {
-                w.locked = false;
-                w.stream_handle
-            }
-            None => return f64::from_bits(TAG_UNDEFINED),
-        }
-    };
-    if let Some(s) = WRITABLE_STREAMS.lock().unwrap().get_mut(&stream_id) {
-        s.writer_handle = None;
-    }
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_closed(writer_handle: f64) -> *mut Promise {
-    let writer_id = writer_handle as usize;
-    match WRITERS.lock().unwrap().get(&writer_id) {
-        Some(w) => w.closed_promise,
-        None => {
-            let p = js_promise_new();
-            js_promise_resolve(p, f64::from_bits(TAG_UNDEFINED));
-            p
-        }
-    }
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_writer_ready(writer_handle: f64) -> *mut Promise {
-    let writer_id = writer_handle as usize;
-    match WRITERS.lock().unwrap().get(&writer_id) {
-        Some(w) => w.ready_promise,
-        None => {
-            let p = js_promise_new();
-            js_promise_resolve(p, f64::from_bits(TAG_UNDEFINED));
-            p
-        }
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────
-// TransformStream FFI
-// ─────────────────────────────────────────────────────────────────────
-
-#[no_mangle]
-pub unsafe extern "C" fn js_transform_stream_new(
-    start_bits: f64,
-    transform_bits: f64,
-    flush_bits: f64,
-    hwm: f64,
-) -> f64 {
-    let start_cb = closure_from_bits(start_bits.to_bits());
-    let transform_cb = closure_from_bits(transform_bits.to_bits());
-    let flush_cb = closure_from_bits(flush_bits.to_bits());
-    alloc_transform_stream(start_cb, transform_cb, flush_cb, None, hwm)
-}
-
-unsafe fn alloc_transform_stream(
-    start_cb: i64,
-    transform_cb: i64,
-    flush_cb: i64,
-    native: Option<NativeTransformKind>,
-    hwm: f64,
-) -> f64 {
-    ensure_gc_registered();
-
-    // Allocate the readable side empty (controller is its own handle).
-    let readable_id = alloc_readable(0, 0, 0, hwm);
-    {
-        let mut g = READABLE_STREAMS.lock().unwrap();
-        if let Some(s) = g.get_mut(&readable_id) {
-            s.started = true;
-        }
-    }
-
-    // Allocate writable side; its write_cb is synthesized via the
-    // dispatcher table below to invoke transform(chunk, controller).
-    let writable_id = next_id(&NEXT_STREAM_ID);
-    let ready = js_promise_new();
-    let closed = js_promise_new();
-    js_promise_resolve(ready, f64::from_bits(TAG_UNDEFINED));
-    WRITABLE_STREAMS.lock().unwrap().insert(
-        writable_id,
-        WritableStreamData {
-            state: WritableState::Writable,
-            // Sentinel: write_cb=0, close_cb=0 — the dispatcher checks
-            // TRANSFORM_PAIRS first and routes through the user transform_cb /
-            // flush_cb instead.
-            write_cb: 0,
-            close_cb: 0,
-            abort_cb: 0,
-            write_queue: VecDeque::new(),
-            in_flight: false,
-            high_water_mark: if hwm.is_nan() || hwm <= 0.0 { 1.0 } else { hwm },
-            writer_handle: None,
-            error_value: 0,
-            ready_promise: ready,
-            closed_promise: closed,
-            close_request_promise: std::ptr::null_mut(),
-            close_started: false,
-        },
-    );
-
-    let id = next_id(&NEXT_STREAM_ID);
-    TRANSFORM_STREAMS.lock().unwrap().insert(
-        id,
-        TransformStreamData {
-            readable_handle: readable_id,
-            writable_handle: writable_id,
-            transform_cb,
-            flush_cb,
-            native,
-        },
-    );
-    TRANSFORM_PAIRS.lock().unwrap().insert(writable_id, id);
-
-    // #1644: TransformStream `start(controller)` fires synchronously at
-    // construction. The controller is the readable-side handle (same value the
-    // transform/flush callbacks receive), so `controller.enqueue(c)` /
-    // `controller.terminate()` / `controller.error(e)` act on the readable.
-    if start_cb != 0 {
-        js_closure_call1(start_cb as *const ClosureHeader, readable_id as f64);
-    }
-    id as f64
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_transform_stream_new_from_transformer_object(
-    transformer: f64,
-    hwm: f64,
-) -> f64 {
-    ensure_gc_registered();
-    js_transform_stream_new(
-        stream_object_field(transformer, b"start"),
-        stream_object_field(transformer, b"transform"),
-        stream_object_field(transformer, b"flush"),
-        hwm,
-    )
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_transform_stream_readable(handle: f64) -> f64 {
-    let id = handle as usize;
-    TRANSFORM_STREAMS
-        .lock()
-        .unwrap()
-        .get(&id)
-        .map(|t| t.readable_handle as f64)
-        .unwrap_or(f64::from_bits(TAG_UNDEFINED))
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_transform_stream_writable(handle: f64) -> f64 {
-    let id = handle as usize;
-    TRANSFORM_STREAMS
-        .lock()
-        .unwrap()
-        .get(&id)
-        .map(|t| t.writable_handle as f64)
-        .unwrap_or(f64::from_bits(TAG_UNDEFINED))
-}
-
-lazy_static::lazy_static! {
-    static ref TRANSFORM_PAIRS: Mutex<HashMap<usize, usize>> = Mutex::new(HashMap::new());
-}
-
-fn transform_writable_for_readable(readable_id: usize) -> Option<usize> {
-    TRANSFORM_STREAMS
-        .lock()
-        .unwrap()
-        .values()
-        .find_map(|t| (t.readable_handle == readable_id).then_some(t.writable_handle))
-}
-
-/// Replacement `writer.write` for the writable side of a TransformStream
-/// — invokes the user transform with (chunk, transformController) where
-/// the transformController is the readable-side stream handle (so
-/// `controller.enqueue(...)` reuses the readable controller path).
-pub(super) unsafe fn transform_write(writable_id: usize, chunk: f64) -> *mut Promise {
-    let promise = js_promise_new();
-    {
-        let g = WRITABLE_STREAMS.lock().unwrap();
-        match g.get(&writable_id) {
-            Some(s) if s.state == WritableState::Writable => {}
-            Some(s) if s.state == WritableState::Errored => {
-                js_promise_reject(promise, f64::from_bits(s.error_value));
-                return promise;
-            }
-            _ => {
-                let err = make_error_with_message("Stream is closed or closing");
-                js_promise_reject(promise, f64::from_bits(err));
-                return promise;
-            }
-        }
-    }
-    let mut handled_native = false;
-    let mut native_error = None;
-    let (transform_cb, readable_id) = {
-        let pairs = TRANSFORM_PAIRS.lock().unwrap();
-        match pairs.get(&writable_id) {
-            Some(t_id) => {
-                let mut g = TRANSFORM_STREAMS.lock().unwrap();
-                match g.get_mut(t_id) {
-                    Some(t) => {
-                        if let Some(native) = t.native.as_mut() {
-                            handled_native = true;
-                            if let Err(error_bits) =
-                                native_transform_write(native, t.readable_handle, chunk)
-                            {
-                                native_error = Some(error_bits);
-                            }
-                        }
-                        (t.transform_cb, t.readable_handle)
-                    }
-                    None => (0, 0),
-                }
-            }
-            None => (0, 0),
-        }
-    };
-    if let Some(error_bits) = native_error {
-        if readable_id != 0 {
-            js_readable_stream_controller_error(readable_id as f64, f64::from_bits(error_bits));
-        }
-        js_promise_reject(promise, f64::from_bits(error_bits));
-        return promise;
-    }
-    if handled_native {
-        js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-        return promise;
-    }
-    if transform_cb != 0 && readable_id != 0 {
-        js_closure_call2(
-            transform_cb as *const ClosureHeader,
-            chunk,
-            readable_id as f64,
-        );
-    } else {
-        // Identity transform — pass-through.
-        js_readable_stream_controller_enqueue(readable_id as f64, chunk);
-    }
-    js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-    promise
-}
-
-pub(super) unsafe fn transform_close(writable_id: usize) -> *mut Promise {
-    let promise = js_promise_new();
-    let mut handled_native = false;
-    let mut native_error = None;
-    let (flush_cb, readable_id) = {
-        let pairs = TRANSFORM_PAIRS.lock().unwrap();
-        match pairs.get(&writable_id) {
-            Some(t_id) => {
-                let mut g = TRANSFORM_STREAMS.lock().unwrap();
-                match g.get_mut(t_id) {
-                    Some(t) => {
-                        if let Some(native) = t.native.as_mut() {
-                            handled_native = true;
-                            if let Err(error_bits) =
-                                native_transform_close(native, t.readable_handle)
-                            {
-                                native_error = Some(error_bits);
-                            }
-                        }
-                        (t.flush_cb, t.readable_handle)
-                    }
-                    None => (0, 0),
-                }
-            }
-            None => (0, 0),
-        }
-    };
-    if let Some(error_bits) = native_error {
-        if readable_id != 0 {
-            js_readable_stream_controller_error(readable_id as f64, f64::from_bits(error_bits));
-        }
-        if let Some(s) = WRITABLE_STREAMS.lock().unwrap().get_mut(&writable_id) {
-            s.state = WritableState::Errored;
-            s.error_value = error_bits;
-            let cp = s.closed_promise;
-            js_promise_reject(cp, f64::from_bits(error_bits));
-        }
-        js_promise_reject(promise, f64::from_bits(error_bits));
-        return promise;
-    }
-    if !handled_native && flush_cb != 0 && readable_id != 0 {
-        js_closure_call1(flush_cb as *const ClosureHeader, readable_id as f64);
-    }
-    if readable_id != 0 {
-        js_readable_stream_controller_close(readable_id as f64);
-    }
-    if let Some(s) = WRITABLE_STREAMS.lock().unwrap().get_mut(&writable_id) {
-        s.state = WritableState::Closed;
-        let cp = s.closed_promise;
-        js_promise_resolve(cp, f64::from_bits(TAG_UNDEFINED));
-    }
-    js_promise_resolve(promise, f64::from_bits(TAG_UNDEFINED));
-    promise
-}
-
-fn split_utf8_prefix(bytes: &[u8]) -> Result<(usize, bool), ()> {
-    match std::str::from_utf8(bytes) {
-        Ok(_) => Ok((bytes.len(), false)),
-        Err(err) => {
-            if err.error_len().is_none() {
-                Ok((err.valid_up_to(), true))
-            } else {
-                Err(())
-            }
-        }
-    }
-}
-
-unsafe fn enqueue_string(readable_id: usize, text: &str) {
-    let ptr = js_string_from_bytes(text.as_ptr(), text.len() as u32);
-    js_readable_stream_controller_enqueue(
-        readable_id as f64,
-        f64::from_bits(JSValue::string_ptr(ptr).bits()),
-    );
-}
-
-unsafe fn native_text_decoder_drain(
-    pending: &mut Vec<u8>,
-    fatal: bool,
-    readable_id: usize,
-    flush: bool,
-) -> Result<(), u64> {
-    if pending.is_empty() {
-        return Ok(());
-    }
-    if fatal {
-        let (valid_len, incomplete) = split_utf8_prefix(pending).map_err(|_| {
-            make_type_error_with_code(
-                "The encoded data was not valid for encoding utf-8",
-                "ERR_ENCODING_INVALID_ENCODED_DATA",
-            )
-        })?;
-        if valid_len > 0 {
-            let text = std::str::from_utf8(&pending[..valid_len]).map_err(|_| {
-                make_type_error_with_code(
-                    "The encoded data was not valid for encoding utf-8",
-                    "ERR_ENCODING_INVALID_ENCODED_DATA",
-                )
-            })?;
-            enqueue_string(readable_id, text);
-            pending.drain(..valid_len);
-        }
-        if flush && !pending.is_empty() {
-            return Err(make_type_error_with_code(
-                "The encoded data was not valid for encoding utf-8",
-                "ERR_ENCODING_INVALID_ENCODED_DATA",
-            ));
-        }
-        if !incomplete && !pending.is_empty() {
-            return Err(make_type_error_with_code(
-                "The encoded data was not valid for encoding utf-8",
-                "ERR_ENCODING_INVALID_ENCODED_DATA",
-            ));
-        }
-        return Ok(());
-    }
-
-    if flush {
-        let text = String::from_utf8_lossy(pending).into_owned();
-        pending.clear();
-        if !text.is_empty() {
-            enqueue_string(readable_id, &text);
-        }
-        return Ok(());
-    }
-
-    let (valid_len, incomplete) = split_utf8_prefix(pending).unwrap_or((pending.len(), false));
-    let emit_len = if incomplete { valid_len } else { pending.len() };
-    if emit_len > 0 {
-        let text = String::from_utf8_lossy(&pending[..emit_len]).into_owned();
-        enqueue_string(readable_id, &text);
-        pending.drain(..emit_len);
-    }
-    Ok(())
-}
-
-fn run_web_compression_codec(
-    format: WebCompressionFormat,
-    decompress: bool,
-    input: &[u8],
-) -> std::io::Result<Vec<u8>> {
-    let mut out = Vec::new();
-    match (format, decompress) {
-        (WebCompressionFormat::Gzip, false) => {
-            GzEncoder::new(input, Compression::default()).read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::Gzip, true) => {
-            GzDecoder::new(input).read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::Deflate, false) => {
-            ZlibEncoder::new(input, Compression::default()).read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::Deflate, true) => {
-            ZlibDecoder::new(input).read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::DeflateRaw, false) => {
-            DeflateEncoder::new(input, Compression::default()).read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::DeflateRaw, true) => {
-            DeflateDecoder::new(input).read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::Brotli, false) => {
-            let mut reader = brotli::CompressorReader::new(input, 4096, 11, 22);
-            reader.read_to_end(&mut out)?;
-        }
-        (WebCompressionFormat::Brotli, true) => {
-            let mut reader = brotli::Decompressor::new(input, 4096);
-            reader.read_to_end(&mut out)?;
-        }
-    }
-    Ok(out)
-}
-
-unsafe fn native_transform_write(
-    native: &mut NativeTransformKind,
-    readable_id: usize,
-    chunk: f64,
-) -> Result<(), u64> {
-    match native {
-        NativeTransformKind::TextEncoder => {
-            let text = js_string_value_to_string(chunk, true).unwrap_or_default();
-            let bytes = alloc_uint8array_from_bytes(text.as_bytes());
-            js_readable_stream_controller_enqueue(readable_id as f64, f64::from_bits(bytes));
-            Ok(())
-        }
-        NativeTransformKind::TextDecoder { fatal, pending } => {
-            let bytes = read_bytes_from_chunk(chunk.to_bits()).ok_or_else(|| {
-                make_type_error_with_code(
-                    "The \"chunk\" argument must be an instance of Buffer, TypedArray, DataView, or ArrayBuffer",
-                    "ERR_INVALID_ARG_TYPE",
-                )
-            })?;
-            pending.extend_from_slice(&bytes);
-            native_text_decoder_drain(pending, *fatal, readable_id, false)
-        }
-        NativeTransformKind::Compression { chunks, .. } => {
-            let bytes = read_bytes_from_chunk(chunk.to_bits()).ok_or_else(|| {
-                make_type_error_with_code(
-                    "The \"chunk\" argument must be an instance of Buffer, TypedArray, DataView, or ArrayBuffer",
-                    "ERR_INVALID_ARG_TYPE",
-                )
-            })?;
-            chunks.extend_from_slice(&bytes);
-            Ok(())
-        }
-    }
-}
-
-unsafe fn native_transform_close(
-    native: &mut NativeTransformKind,
-    readable_id: usize,
-) -> Result<(), u64> {
-    match native {
-        NativeTransformKind::TextEncoder => Ok(()),
-        NativeTransformKind::TextDecoder { fatal, pending } => {
-            native_text_decoder_drain(pending, *fatal, readable_id, true)
-        }
-        NativeTransformKind::Compression {
-            format,
-            decompress,
-            chunks,
-        } => match run_web_compression_codec(*format, *decompress, chunks) {
-            Ok(out) => {
-                let chunk = alloc_uint8array_from_bytes(&out);
-                js_readable_stream_controller_enqueue(readable_id as f64, f64::from_bits(chunk));
-                chunks.clear();
-                Ok(())
-            }
-            Err(err) => Err(make_type_error_with_code(&err.to_string(), "Z_DATA_ERROR")),
-        },
-    }
-}
-
-unsafe fn attach_stream_field(object_value: f64, name: &[u8], value: f64) {
-    let ptr = js_nanbox_get_pointer(object_value) as *mut ObjectHeader;
-    if ptr.is_null() || (ptr as usize) < 0x10000 {
-        return;
-    }
-    let key = js_string_from_bytes(name.as_ptr(), name.len() as u32);
-    js_object_set_field_by_name(ptr, key, value);
-}
-
-unsafe fn attach_stream_string_field(object_value: f64, name: &[u8], value: &[u8]) {
-    let ptr = js_string_from_bytes(value.as_ptr(), value.len() as u32);
-    attach_stream_field(
-        object_value,
-        name,
-        f64::from_bits(JSValue::string_ptr(ptr).bits()),
-    );
-}
-
-unsafe fn attach_stream_bool_field(object_value: f64, name: &[u8], value: bool) {
-    attach_stream_field(
-        object_value,
-        name,
-        f64::from_bits(if value { TAG_TRUE } else { TAG_FALSE }),
-    );
-}
-
-unsafe fn attach_transform_endpoints(object_value: f64, readable_id: usize, writable_id: usize) {
-    attach_stream_field(object_value, b"readable", readable_id as f64);
-    attach_stream_field(object_value, b"writable", writable_id as f64);
-}
-
-unsafe fn bool_option(options: f64, name: &[u8]) -> bool {
-    let jsval = JSValue::from_bits(options.to_bits());
-    if !jsval.is_pointer() {
-        return false;
-    }
-    let value =
-        perry_runtime::value::js_get_property(options, name.as_ptr() as i64, name.len() as i64);
-    perry_runtime::value::js_is_truthy(value) != 0
-}
-
-unsafe fn parse_text_decoder_stream_label(label: f64) {
-    let jsval = JSValue::from_bits(label.to_bits());
-    if jsval.is_undefined()
-        || value_string_equals(label, b"utf-8")
-        || value_string_equals(label, b"utf8")
-    {
-        return;
-    }
-    let label_text = js_string_value_to_string(label, true).unwrap_or_default();
-    let message = format!("The \"{label_text}\" encoding is not supported");
-    throw_range_error_with_code(&message, "ERR_ENCODING_NOT_SUPPORTED");
-}
-
-unsafe fn parse_web_compression_format(value: f64, constructor_name: &str) -> WebCompressionFormat {
-    if value_string_equals(value, b"gzip") {
-        return WebCompressionFormat::Gzip;
-    }
-    if value_string_equals(value, b"deflate") {
-        return WebCompressionFormat::Deflate;
-    }
-    if value_string_equals(value, b"deflate-raw") {
-        return WebCompressionFormat::DeflateRaw;
-    }
-    if value_string_equals(value, b"brotli") {
-        return WebCompressionFormat::Brotli;
-    }
-    let received =
-        js_string_value_to_string(value, true).unwrap_or_else(|| "undefined".to_string());
-    let message = format!(
-        "Failed to construct '{constructor_name}': 1st argument value '{received}' is not a valid enum value of type CompressionFormat."
-    );
-    throw_type_error_with_code(&message, "ERR_INVALID_ARG_VALUE");
-}
-
-unsafe fn build_native_transform_object(object_value: f64, native: NativeTransformKind) -> f64 {
-    let handle = alloc_transform_stream(0, 0, 0, Some(native), 1.0);
-    let readable_id = js_transform_stream_readable(handle) as usize;
-    let writable_id = js_transform_stream_writable(handle) as usize;
-    attach_transform_endpoints(object_value, readable_id, writable_id);
-    object_value
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_stream_web_text_encoder_stream_new() -> f64 {
-    let object = perry_runtime::object::js_text_encoder_stream_new();
-    attach_stream_string_field(object, b"encoding", b"utf-8");
-    build_native_transform_object(object, NativeTransformKind::TextEncoder)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_stream_web_text_decoder_stream_new(label: f64, options: f64) -> f64 {
-    parse_text_decoder_stream_label(label);
-    let fatal = bool_option(options, b"fatal");
-    let ignore_bom = bool_option(options, b"ignoreBOM");
-    let object = perry_runtime::object::js_text_decoder_stream_new();
-    attach_stream_string_field(object, b"encoding", b"utf-8");
-    attach_stream_bool_field(object, b"fatal", fatal);
-    attach_stream_bool_field(object, b"ignoreBOM", ignore_bom);
-    build_native_transform_object(
-        object,
-        NativeTransformKind::TextDecoder {
-            fatal,
-            pending: Vec::new(),
-        },
-    )
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_stream_web_compression_stream_new(format: f64) -> f64 {
-    let format = parse_web_compression_format(format, "CompressionStream");
-    let object = perry_runtime::object::js_compression_stream_new();
-    build_native_transform_object(
-        object,
-        NativeTransformKind::Compression {
-            format,
-            decompress: false,
-            chunks: Vec::new(),
-        },
-    )
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_stream_web_decompression_stream_new(format: f64) -> f64 {
-    let format = parse_web_compression_format(format, "DecompressionStream");
-    let object = perry_runtime::object::js_decompression_stream_new();
-    build_native_transform_object(
-        object,
-        NativeTransformKind::Compression {
-            format,
-            decompress: true,
-            chunks: Vec::new(),
-        },
-    )
-}
-
-// ─────────────────────────────────────────────────────────────────────
-// Stubs for deferred surface (issue #237 followups)
-// ─────────────────────────────────────────────────────────────────────
-
-#[no_mangle]
-pub unsafe extern "C" fn js_streams_throw_byob_not_implemented() -> f64 {
-    let err = make_error_with_message("BYOB readers are not yet implemented (issue #237 followup)");
-    perry_runtime::exception::js_throw(f64::from_bits(err));
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_streams_throw_byte_length_not_implemented() -> f64 {
-    let err = make_error_with_message(
-        "ByteLengthQueuingStrategy is not yet implemented (issue #237 followup)",
-    );
-    perry_runtime::exception::js_throw(f64::from_bits(err));
-}
-
-// ─────────────────────────────────────────────────────────────────────
-// Public helpers used by other crates / tests
-// ─────────────────────────────────────────────────────────────────────
-
-// ─────────────────────────────────────────────────────────────────────
-// Subclass support (issue #562)
-//
-// User classes extending `WritableStream` / `ReadableStream` /
-// `TransformStream` get an underlying-stream registry handle allocated
-// at `super({ ... })` time and stashed on `this` under the hidden field
-// `__perry_stream_handle__`. The dispatch arms in `lower_call.rs` route
-// the receiver / destination through `js_stream_unwrap_handle` before
-// the FFI call so subclass instances and bare handles are
-// interchangeable.
-// ─────────────────────────────────────────────────────────────────────
-
-/// Hidden field name used to stash the underlying-stream registry id on
-/// a subclass instance. Read by `js_stream_unwrap_handle`, written by
-/// the three `*_subclass_init` helpers below.
-const SUBCLASS_HANDLE_FIELD: &[u8] = b"__perry_stream_handle__";
-
-unsafe fn subclass_handle_key() -> *const perry_runtime::StringHeader {
-    js_string_from_bytes(
-        SUBCLASS_HANDLE_FIELD.as_ptr(),
-        SUBCLASS_HANDLE_FIELD.len() as u32,
-    )
-}
-
-unsafe fn this_object_ptr(this_bits: f64) -> Option<*mut ObjectHeader> {
-    let bits = this_bits.to_bits();
-    let top16 = bits >> 48;
-    if top16 != 0x7FFD {
-        return None;
-    }
-    let raw = (bits & POINTER_MASK) as *mut ObjectHeader;
-    if raw.is_null() || (raw as usize) < 0x10000 {
-        return None;
-    }
-    Some(raw)
-}
-
-unsafe fn attach_handle_to_this(this_bits: f64, handle_id: usize) {
-    if let Some(obj) = this_object_ptr(this_bits) {
-        let key = subclass_handle_key();
-        // Stored as plain f64 numeric — same ABI the rest of the stream
-        // FFIs use for handles. `js_stream_unwrap_handle` reads it back.
-        js_object_set_field_by_name(obj, key, handle_id as f64);
-    }
-}
-
-/// Resolve a stream receiver / argument to a numeric registry handle.
-///
-/// For raw numeric handles (the value `js_writable_stream_new` etc.
-/// return) the input is returned unchanged. For NaN-boxed pointer-tagged
-/// JS objects (subclass instances), reads the hidden
-/// `__perry_stream_handle__` field. Falls back to the input when the
-/// field is missing — caller's downstream FFI will then no-op on a
-/// 0-or-bogus handle exactly as it did pre-#562.
-#[no_mangle]
-pub unsafe extern "C" fn js_stream_unwrap_handle(value: f64) -> f64 {
-    let bits = value.to_bits();
-    let top16 = bits >> 48;
-    if top16 != 0x7FFD {
-        return value;
-    }
-    let Some(obj) = this_object_ptr(value) else {
-        return value;
-    };
-    let key = subclass_handle_key();
-    let result = js_object_get_field_by_name(obj, key);
-    let result_bits = result.bits();
-    if result_bits == TAG_UNDEFINED || result_bits == TAG_NULL {
-        return value;
-    }
-    f64::from_bits(result_bits)
-}
-
-#[inline]
-fn box_promise(p: *mut Promise) -> f64 {
-    f64::from_bits(JSValue::pointer(p as *const u8).bits())
-}
-
-/// #1545: probe used by `js_native_call_method` to recognise a numeric receiver
-/// as a live Web Streams handle (readable/writable/reader/writer). Only ids in
-/// the reserved stream range that are present in a registry qualify.
-#[no_mangle]
-pub extern "C" fn js_stream_handle_is_registered(id: usize) -> bool {
-    js_stream_handle_kind(id) != 0
-}
-
-/// #1545: classify a numeric Web Streams handle for `instanceof`, dispatch,
-/// and `Object.prototype.toString` tags.
-/// 0 = not a stream, 1 = ReadableStream, 2 = WritableStream, 3 = reader,
-/// 4 = writer, 5 = TransformStream.
-#[no_mangle]
-pub extern "C" fn js_stream_handle_kind(id: usize) -> u8 {
-    if !(STREAM_HANDLE_ID_START..STREAM_HANDLE_ID_END).contains(&id) {
-        return 0;
-    }
-    if READABLE_STREAMS
-        .lock()
-        .map(|m| m.contains_key(&id))
-        .unwrap_or(false)
-    {
-        return 1;
-    }
-    if WRITABLE_STREAMS
-        .lock()
-        .map(|m| m.contains_key(&id))
-        .unwrap_or(false)
-    {
-        return 2;
-    }
-    if READERS.lock().map(|m| m.contains_key(&id)).unwrap_or(false) {
-        return 3;
-    }
-    if WRITERS.lock().map(|m| m.contains_key(&id)).unwrap_or(false) {
-        return 4;
-    }
-    if TRANSFORM_STREAMS
-        .lock()
-        .map(|m| m.contains_key(&id))
-        .unwrap_or(false)
-    {
-        return 5;
-    }
-    0
-}
-
-/// #1545: runtime method dispatch for Web Streams handles whose static type
-/// the codegen could not track. The static `module == "readable_stream"` /
-/// `"reader"` / … NativeMethodCall arms only fire when the receiver is a local
-/// whose inferred type is the stream class. Chained / member results lose that
-/// type — e.g. `src.pipeThrough(ts).getReader()`, `ts.readable.getReader()`,
-/// `rs.tee()[0].getReader()`, `const r = rs.getReader(); r.read()` — and lower
-/// to a generic method call that reaches `js_native_call_method` →
-/// `js_handle_method_dispatch` with a bare numeric handle.
-///
-/// Because every Web Streams handle now lives in one shared id space based at
-/// `STREAM_HANDLE_ID_START` (see `NEXT_STREAM_ID`), the handle is (a)
-/// recognisable by range and (b) present in exactly one of the five registries,
-/// so routing by
-/// `(registry-membership, method-name)` is unambiguous and can never collide
-/// with another handle subsystem. Returns `None` when the handle isn't a stream
-/// handle or the method isn't a stream method, so the generic dispatcher falls
-/// through to the next arm unchanged.
-pub(crate) unsafe fn dispatch_stream_method(
-    handle: f64,
-    method: &str,
-    args: &[f64],
-) -> Option<f64> {
-    let id = handle as usize;
-    if !(STREAM_HANDLE_ID_START..STREAM_HANDLE_ID_END).contains(&id) {
-        return None;
-    }
-    let arg0 = args
-        .first()
-        .copied()
-        .unwrap_or(f64::from_bits(TAG_UNDEFINED));
-    let arg1 = args
-        .get(1)
-        .copied()
-        .unwrap_or(f64::from_bits(TAG_UNDEFINED));
-
-    // Probe each registry for membership first (dropping the guard before we
-    // call the FFI, which re-locks the same registry).
-    let is_reader = READERS.lock().unwrap().contains_key(&id);
-    if is_reader {
-        match method {
-            "read" => return Some(box_promise(js_reader_read(handle))),
-            "releaseLock" => return Some(js_reader_release_lock(handle)),
-            "cancel" => return Some(box_promise(js_reader_cancel(handle, arg0))),
-            _ => return None,
-        }
-    }
-    let is_writer = WRITERS.lock().unwrap().contains_key(&id);
-    if is_writer {
-        match method {
-            "write" => return Some(box_promise(js_writer_write(handle, arg0))),
-            "close" => return Some(box_promise(js_writer_close(handle))),
-            "abort" => return Some(box_promise(js_writer_abort(handle, arg0))),
-            "releaseLock" => return Some(js_writer_release_lock(handle)),
-            _ => return None,
-        }
-    }
-    let is_readable = READABLE_STREAMS.lock().unwrap().contains_key(&id);
-    if is_readable {
-        match method {
-            "getReader" => return Some(js_readable_stream_get_reader_with_options(handle, arg0)),
-            "values" | "@@asyncIterator" => return Some(js_readable_stream_values(handle)),
-            "cancel" => return Some(box_promise(js_readable_stream_cancel(handle, arg0))),
-            "tee" => return Some(js_readable_stream_tee(handle)),
-            "pipeTo" => return Some(box_promise(js_readable_stream_pipe_to(handle, arg0, arg1))),
-            "pipeThrough" => {
-                let transform = js_stream_unwrap_handle(arg0);
-                let writable = js_transform_stream_writable(transform);
-                let readable = js_transform_stream_readable(transform);
-                return Some(js_readable_stream_pipe_through(handle, writable, readable));
-            }
-            // #1644: a readable handle is also its own controller. The
-            // start/transform/flush callbacks receive it as `controller`, so
-            // `controller.enqueue/close/error/terminate` dispatch here when the
-            // controller param is generically typed. `terminate()` ends the
-            // readable side (TransformStreamDefaultController.terminate).
-            "enqueue" => return Some(js_readable_stream_controller_enqueue(handle, arg0)),
-            "close" | "terminate" => return Some(js_readable_stream_controller_close(handle)),
-            "error" => return Some(js_readable_stream_controller_error(handle, arg0)),
-            _ => return None,
-        }
-    }
-    let is_writable = WRITABLE_STREAMS.lock().unwrap().contains_key(&id);
-    if is_writable {
-        match method {
-            "getWriter" => return Some(js_writable_stream_get_writer(handle)),
-            "abort" => return Some(box_promise(js_writable_stream_abort(handle, arg0))),
-            "close" => return Some(box_promise(js_writable_stream_close(handle))),
-            _ => return None,
-        }
-    }
-    None
-}
-
-/// #1670: property reads on a numeric Web Streams handle that reached the
-/// generic field-get path (e.g. inline `res.body.locked`, where the
-/// intermediate stream id never became a typed local). Returns the WHATWG
-/// getter property value, a bound-method closure for callable members (so
-/// `typeof rs.getReader === "function"` and a subsequent call routes back
-/// through `js_native_call_method`'s #1545 stream branch → `dispatch_stream_method`),
-/// or `undefined` for any other property. Crucially this NEVER dereferences
-/// the float id as a pointer — the pre-#1670 generic field-get segfaulted on
-/// `res.body.locked`. Gated by the caller on stream-registry membership.
-pub(crate) unsafe fn dispatch_stream_property(handle: f64, name: &str) -> f64 {
-    let undefined = f64::from_bits(TAG_UNDEFINED);
-    let id = handle as usize;
-    // Kind: 1=ReadableStream, 2=WritableStream, 3=reader, 4=writer.
-    let kind = js_stream_handle_kind(id);
-    if kind == 0 {
-        return undefined;
-    }
-    // WHATWG getter properties (the rest fall through to bound-method /
-    // undefined). `locked` is the one #1670 exercises (`res.body.locked`).
-    match (kind, name) {
-        (1, "locked") => return js_readable_stream_locked(handle),
-        (2, "locked") => return js_writable_stream_locked(handle),
-        (3, "closed") => return box_promise(js_reader_closed(handle)),
-        _ => {}
-    }
-    // Callable members → bound-method closure so `typeof` reports
-    // "function". The name must be a `&'static [u8]` because
-    // `js_class_method_bind` stores the raw pointer in the closure.
-    // The receiver is the raw float handle (not NaN-boxed) so that when the
-    // bound method is called, `js_native_call_method`'s stream branch fires.
-    let method: Option<&'static [u8]> = match (kind, name) {
-        (1, "getReader") => Some(b"getReader"),
-        (1, "cancel") => Some(b"cancel"),
-        (1, "tee") => Some(b"tee"),
-        (1, "pipeTo") => Some(b"pipeTo"),
-        (1, "pipeThrough") => Some(b"pipeThrough"),
-        (2, "getWriter") => Some(b"getWriter"),
-        (2, "abort") => Some(b"abort"),
-        (2, "close") => Some(b"close"),
-        (3, "read") => Some(b"read"),
-        (3, "releaseLock") => Some(b"releaseLock"),
-        (3, "cancel") => Some(b"cancel"),
-        (4, "write") => Some(b"write"),
-        (4, "close") => Some(b"close"),
-        (4, "abort") => Some(b"abort"),
-        (4, "releaseLock") => Some(b"releaseLock"),
-        _ => None,
-    };
-    if let Some(name_bytes) = method {
-        extern "C" {
-            fn js_class_method_bind(
-                instance: f64,
-                method_name_ptr: *const u8,
-                method_name_len: usize,
-            ) -> f64;
-        }
-        return js_class_method_bind(handle, name_bytes.as_ptr(), name_bytes.len());
-    }
-    undefined
-}
-
-/// `super({ start, pull, cancel })` dispatch for `class X extends ReadableStream`.
-/// Allocates the underlying readable handle, stashes it on `this`, runs
-/// the user `start` callback synchronously (mirrors `js_readable_stream_new`).
-#[no_mangle]
-pub unsafe extern "C" fn js_readable_stream_subclass_init(
-    this_bits: f64,
-    start_bits: f64,
-    pull_bits: f64,
-    cancel_bits: f64,
-    hwm: f64,
-) -> f64 {
-    ensure_gc_registered();
-    let id = alloc_readable(
-        closure_from_bits(start_bits.to_bits()),
-        closure_from_bits(pull_bits.to_bits()),
-        closure_from_bits(cancel_bits.to_bits()),
-        hwm,
-    );
-    attach_handle_to_this(this_bits, id);
-    invoke_start(id);
-    maybe_pull(id);
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-/// `super({ write, close, abort })` dispatch for `class X extends WritableStream`.
-#[no_mangle]
-pub unsafe extern "C" fn js_writable_stream_subclass_init(
-    this_bits: f64,
-    write_bits: f64,
-    close_bits: f64,
-    abort_bits: f64,
-    hwm: f64,
-) -> f64 {
-    ensure_gc_registered();
-    let id = alloc_writable(
-        closure_from_bits(write_bits.to_bits()),
-        closure_from_bits(close_bits.to_bits()),
-        closure_from_bits(abort_bits.to_bits()),
-        hwm,
-    );
-    attach_handle_to_this(this_bits, id);
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-/// `super({ transform, flush })` dispatch for `class X extends TransformStream`.
-/// Allocates the transform-stream pair (readable + writable + the
-/// dispatcher row in `TRANSFORM_PAIRS`) — same shape as
-/// `js_transform_stream_new` — and stashes the transform handle id on
-/// `this`. `pipeThrough(subclass)` then calls `js_transform_stream_writable`
-/// / `_readable` after `js_stream_unwrap_handle`, finding the same
-/// readable / writable sub-handles.
-#[no_mangle]
-pub unsafe extern "C" fn js_transform_stream_subclass_init(
-    this_bits: f64,
-    transform_bits: f64,
-    flush_bits: f64,
-    hwm: f64,
-) -> f64 {
-    // #1644: subclass `super({...})` doesn't thread a `start` hook through this
-    // path (the #562 subclass shim only forwards transform/flush); pass undefined.
-    let handle = js_transform_stream_new(
-        f64::from_bits(TAG_UNDEFINED),
-        transform_bits,
-        flush_bits,
-        hwm,
-    );
-    attach_handle_to_this(this_bits, handle as usize);
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-/// Read every queued chunk into a Vec<u8>, draining the stream. Used by
-/// `new Response(stream)` / `new Request(url, { body: stream })` — we
-/// drain the buffered chunks at construction time so the resulting
-/// Response.body bytes match what a real serializer would produce.
-#[doc(hidden)]
-pub fn drain_readable_into_bytes(stream_id: usize) -> Vec<u8> {
-    let mut out = Vec::new();
-    let chunks: Vec<u64> = {
-        let mut g = READABLE_STREAMS.lock().unwrap();
-        match g.get_mut(&stream_id) {
-            Some(s) => {
-                let drained: Vec<u64> = s.chunks.drain(..).collect();
-                s.state = ReadableState::Closed;
-                drained
-            }
-            None => return out,
-        }
-    };
-    for chunk in chunks {
-        unsafe {
-            if let Some(bytes) = read_bytes_from_chunk(chunk) {
-                out.extend_from_slice(&bytes);
-            }
-        }
-    }
-    out
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn stream_ids_live_outside_pointer_tag_small_handle_band() {
-        let id = next_id(&NEXT_STREAM_ID);
-        assert!(
-            (STREAM_HANDLE_ID_START..STREAM_HANDLE_ID_END).contains(&id),
-            "stream id {id:#x} must stay in the raw numeric stream band"
-        );
-        assert!(
-            id >= 0x100000,
-            "stream id {id:#x} must not overlap pointer-tagged small handles"
-        );
-    }
-
-    #[test]
-    fn root_scanner_emits_callbacks_chunks_and_promises() {
-        {
-            let mut readable = READABLE_STREAMS.lock().unwrap();
-            readable.clear();
-            readable.insert(
-                1,
-                ReadableStreamData {
-                    state: ReadableState::Errored,
-                    chunks: VecDeque::from([0x7FFD_0000_0000_1234]),
-                    pending_reads: VecDeque::from([0x2345_6780 as *mut Promise]),
-                    start_cb: 0x3456_7890,
-                    pull_cb: 0,
-                    cancel_cb: 0,
-                    high_water_mark: 1.0,
-                    strategy_size_cb: 0,
-                    is_byte_stream: false,
-                    pull_returns_byte_chunk: false,
-                    pulling: false,
-                    started: false,
-                    reader_handle: None,
-                    error_value: 0x7FFF_0000_0000_4567,
-                    pending_error_after_chunks: None,
-                    canceled: false,
-                },
-            );
-        }
-
-        let mut emitted = Vec::new();
-        scan_stream_roots(&mut |value| emitted.push(value.to_bits()));
-
-        assert!(emitted.contains(&0x7FFD_0000_0000_1234));
-        assert!(emitted.contains(&(0x7FFD_0000_0000_0000 | 0x2345_6780)));
-        assert!(emitted.contains(&(0x7FFD_0000_0000_0000 | 0x3456_7890)));
-        assert!(emitted.contains(&0x7FFF_0000_0000_4567));
-        READABLE_STREAMS.lock().unwrap().clear();
-    }
-
-    #[test]
-    fn web_compression_formats_round_trip() {
-        let input = b"hello stream/web compression";
-        for format in [
-            WebCompressionFormat::Gzip,
-            WebCompressionFormat::Deflate,
-            WebCompressionFormat::DeflateRaw,
-            WebCompressionFormat::Brotli,
-        ] {
-            let compressed = run_web_compression_codec(format, false, input).unwrap();
-            assert!(!compressed.is_empty());
-            let decompressed = run_web_compression_codec(format, true, &compressed).unwrap();
-            assert_eq!(decompressed, input);
-        }
-    }
-
-    #[test]
-    fn utf8_split_prefix_tracks_incomplete_sequence() {
-        assert_eq!(split_utf8_prefix(&[0x68, 0xc3]).unwrap(), (1, true));
-        assert_eq!(split_utf8_prefix(&[0xc3, 0xa9]).unwrap(), (2, false));
-        assert!(split_utf8_prefix(&[0xff]).is_err());
-    }
 }
